Refactor app.ts (entrypoint) and oclif-specific files
Change-type: patch
