+++ lib/app.ts
@@ -14,147 +14,103 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { stripIndent } from 'common-tags';
-
-import { exitWithExpectedError } from './utils/patterns';
-
-export interface AppOptions {
-	// Prevent the default behaviour of flushing stdout after running a command
-	noFlush?: boolean;
-}
 
 /**
- * Simple command-line pre-parsing to choose between oclif or Capitano.
- * @param argv process.argv
+ * Sentry.io setup
+ * @see https://docs.sentry.io/clients/node/
  */
-function routeCliFramework(argv: string[], options: AppOptions): void {
-	if (process.env.DEBUG) {
-		console.log(
-			`[debug] original argv0="${process.argv0}" argv=[${argv}] length=${
-				argv.length
-			}`,
-		);
-	}
-	const cmdSlice = argv.slice(2);
+function setupRaven() {
+	const Raven = require('raven');
+	Raven.disableConsoleAlerts();
+	Raven.config(require('./config').sentryDsn, {
+		captureUnhandledRejections: true,
+		autoBreadcrumbs: true,
+		release: require('../package.json').version,
+	}).install(function(_logged: any, error: Error) {
+		console.error(error);
+		return process.exit(1);
+	});
 
-	// Look for commands that have been deleted, to print a notice
-	checkDeletedCommand(cmdSlice);
+	Raven.setContext({
+		extra: {
+			args: process.argv,
+			node_version: process.version,
+		},
+	});
+}
 
-	if (cmdSlice.length > 0) {
-		// convert 'balena --version' or 'balena -v' to 'balena version'
-		if (['--version', '-v'].includes(cmdSlice[0])) {
-			cmdSlice[0] = 'version';
-		}
-		// convert 'balena --help' or 'balena -h' to 'balena help'
-		else if (['--help', '-h'].includes(cmdSlice[0])) {
-			cmdSlice[0] = 'help';
-		}
-		// convert e.g. 'balena help env add' to 'balena env add --help'
-		if (cmdSlice.length > 1 && cmdSlice[0] === 'help') {
-			cmdSlice.shift();
-			cmdSlice.push('--help');
-		}
-	}
-
-	const [isOclif, isTopic] = isOclifCommand(cmdSlice);
-
-	if (isOclif) {
-		let oclifArgs = cmdSlice;
-		if (isTopic) {
-			// convert space-separated commands to oclif's topic:command syntax
-			oclifArgs = [cmdSlice[0] + ':' + cmdSlice[1], ...cmdSlice.slice(2)];
-		}
-		if (process.env.DEBUG) {
-			console.log(
-				`[debug] new argv=[${[
-					argv[0],
-					argv[1],
-					...oclifArgs,
-				]}] length=${oclifArgs.length + 2}`,
-			);
-		}
-		return require('./app-oclif').run(oclifArgs, options);
-	} else {
-		return require('./app-capitano').run(argv);
+function checkNodeVersion() {
+	const validNodeVersions = require('../package.json').engines.node;
+	if (!require('semver').satisfies(process.version, validNodeVersions)) {
+		const { stripIndent } = require('common-tags');
+		console.warn(stripIndent`
+			------------------------------------------------------------------------------
+			Warning: Node version "${
+				process.version
+			}" does not match required versions "${validNodeVersions}".
+			This may cause unexpected behaviour. To upgrade Node, visit:
+			https://nodejs.org/en/download/
+			------------------------------------------------------------------------------
+			`);
 	}
 }
 
-/**
- *
- * @param argvSlice process.argv.slice(2)
- */
-function checkDeletedCommand(argvSlice: string[]): void {
-	if (argvSlice[0] === 'help') {
-		argvSlice = argvSlice.slice(1);
+function setupGlobalHttpProxy() {
+	// Doing this before requiring any other modules,
+	// including the 'balena-sdk', to prevent any module from reading the http proxy config
+	// before us
+	const globalTunnel = require('global-tunnel-ng');
+	const settings = require('balena-settings-client');
+	let proxy;
+	try {
+		proxy = settings.get('proxy') || null;
+	} catch (error1) {
+		proxy = null;
 	}
-	function replaced(
-		oldCmd: string,
-		alternative: string,
-		version: string,
-		verb = 'replaced',
-	) {
-		exitWithExpectedError(stripIndent`
-			Note: the command "balena ${oldCmd}" was ${verb} in CLI version ${version}.
-			Please use "balena ${alternative}" instead.
-		`);
-	}
-	function removed(oldCmd: string, alternative: string, version: string) {
-		let msg = `Note: the command "balena ${oldCmd}" was removed in CLI version ${version}.`;
-		if (alternative) {
-			msg = [msg, alternative].join('\n');
-		}
-		exitWithExpectedError(msg);
-	}
-	const stopAlternative =
-		'Please use "balena ssh -s" to access the host OS, then use `balena-engine stop`.';
-	const cmds: { [cmd: string]: [(...args: any) => void, ...string[]] } = {
-		sync: [replaced, 'push', 'v11.0.0', 'removed'],
-		'local logs': [replaced, 'logs', 'v11.0.0'],
-		'local push': [replaced, 'push', 'v11.0.0'],
-		'local scan': [replaced, 'scan', 'v11.0.0'],
-		'local ssh': [replaced, 'ssh', 'v11.0.0'],
-		'local stop': [removed, stopAlternative, 'v11.0.0'],
-	};
-	let cmd: string | undefined;
-	if (argvSlice.length > 1) {
-		cmd = [argvSlice[0], argvSlice[1]].join(' ');
-	} else if (argvSlice.length > 0) {
-		cmd = argvSlice[0];
-	}
-	if (cmd && Object.getOwnPropertyNames(cmds).includes(cmd)) {
-		cmds[cmd][0](cmd, ...cmds[cmd].slice(1));
-	}
+
+	// Init the tunnel even if the proxy is not configured
+	// because it can also get the proxy from the http(s)_proxy env var
+	// If that is not set as well the initialize will do nothing
+	globalTunnel.initialize(proxy);
+
+	// TODO: make this a feature of capitano https://github.com/balena-io/capitano/issues/48
+	(global as any).PROXY_CONFIG = globalTunnel.proxyConfig;
 }
 
-/**
- * Determine whether the CLI command has been converted from Capitano to oclif.
- * Return an array of two boolean values:
- *   r[0] : whether the CLI command is implemented with oclif
- *   r[1] : if r[0] is true, whether the CLI command is implemented with
- *          oclif "topics" (colon-separated subcommands like `env:add`)
- * @param argvSlice process.argv.slice(2)
- */
-function isOclifCommand(argvSlice: string[]): [boolean, boolean] {
-	// Look for commands that have been transitioned to oclif
-	if (argvSlice.length > 0) {
-		// balena version
-		if (argvSlice[0] === 'version') {
-			return [true, false];
-		}
-		if (argvSlice.length > 1) {
-			// balena env add
-			if (argvSlice[0] === 'env' && argvSlice[1] === 'add') {
-				return [true, true];
-			}
-
-			// balena env rm
-			if (argvSlice[0] === 'env' && argvSlice[1] === 'rm') {
-				return [true, true];
-			}
-		}
-	}
-	return [false, false];
+function setupBalenaSdkSharedOptions() {
+	// We don't yet use balena-sdk directly everywhere, but we set up shared
+	// options correctly so we can do safely in submodules
+	const BalenaSdk = require('balena-sdk');
+	const settings = require('balena-settings-client');
+	BalenaSdk.setSharedOptions({
+		apiUrl: settings.get('apiUrl'),
+		imageMakerUrl: settings.get('imageMakerUrl'),
+		dataDirectory: settings.get('dataDirectory'),
+		retries: 2,
+	});
 }
+
+function globalInit() {
+	setupRaven();
+	checkNodeVersion();
+	setupGlobalHttpProxy();
+	setupBalenaSdkSharedOptions();
+
+	// Assign bluebird as the global promise library.
+	// stream-to-promise will produce native promises if not for this module,
+	// which is likely to lead to errors as much of the CLI coffeescript code
+	// expects bluebird promises.
+	// The registration is only run if it hasn't already happened (for example
+	// in a test case).
+	if (!(global as any)['@@any-promise/REGISTRATION']) {
+		require('any-promise/register/bluebird');
+	}
+
+	// check for CLI updates once a day
+	require('./utils/update').notify();
+}
+
+type AppOptions = import('./preparser').AppOptions;
 
 /**
  * CLI entrypoint, but see also `bin/balena` and `bin/balena-dev` which
@@ -164,6 +120,6 @@ export function run(cliArgs = process.argv, options: AppOptions = {}): void {
 	// globalInit() must be called very early on (before other imports) because
 	// it sets up Sentry error reporting, global HTTP proxy settings, balena-sdk
 	// shared options, and performs node version requirement checks.
-	require('./app-common').globalInit();
-	return routeCliFramework(cliArgs, options);
+	globalInit();
+	return require('./preparser').routeCliFramework(cliArgs, options);
 }
