+++ lib/oclif/hooks/prerun/track.ts
@@ -31,7 +31,7 @@ export let trackPromise: PromiseLike<void>;
  * literally so: 'NAME' and 'VALUE' are NOT replaced with actual values.
  */
 const hook: Hook<'prerun'> = async function(options) {
-	const events = await import('../../events');
+	const events = await import('../../../events');
 	const usage: string | string[] | undefined = options.Command.usage;
 	const cmdSignature =
 		usage == null ? '*' : typeof usage === 'string' ? usage : usage.join(' ');
