+++ package.json
@@ -2,7 +2,7 @@
   "name": "balena-cli",
   "version": "11.9.1",
   "description": "The official balena CLI tool",
-  "main": "./build/actions/index.js",
+  "main": "./build/app.js",
   "homepage": "https://github.com/balena-io/balena-cli",
   "repository": {
     "type": "git",
@@ -29,7 +29,7 @@
       "node_modules/raven/lib/instrumentation/*.js"
     ],
     "assets": [
-      "build/actions-oclif",
+      "build/oclif/**/*",
       "build/auth/pages/*.ejs",
       "node_modules/resin-discoverable-services/services/**/*"
     ]
@@ -70,9 +70,9 @@
   },
   "oclif": {
     "bin": "balena",
-    "commands": "./build/actions-oclif",
+    "commands": "./build/oclif/commands",
     "hooks": {
-      "prerun": "./build/hooks/prerun/track"
+      "prerun": "./build/oclif/hooks/prerun/track"
     },
     "macos": {
       "identifier": "io.balena.cli",
