+++ typings/hidepath/index.d.ts
@@ -1,0 +1,7 @@
+declare module 'hidepath' {
+	type HidepathFn = (path: string) => string;
+
+	const hidepath: HidepathFn;
+
+	export = hidepath;
+}
