+++ typings/home-or-tmp/index.d.ts
@@ -1,0 +1,5 @@
+declare module 'home-or-tmp' {
+	const homeOrTmp: string;
+
+	export = homeOrTmp;
+}
